reverted type changes

+++ BetterNumbers/src/linalg/Matrix2D.cpp
@@ -11,7 +11,7 @@ static unsigned index_;
 
 // CONSTRUCTORS
 
-Matrix2D::Matrix2D(vector *data, int rows, int cols) {
+Matrix2D::Matrix2D(vector *data, unsigned short rows, unsigned short cols) {
 
 	if (areValidParams(data, rows, cols)) {
 		__rows = rows;
@@ -35,27 +35,27 @@ Matrix2D::Matrix2D(nvector *data) {
 	}
 }
 
-Matrix2D Matrix2D::zeros(int _l) {
+Matrix2D Matrix2D::zeros(unsigned short _l) {
 	return zeros(_l, _l);
 }
 
-Matrix2D Matrix2D::zeros(int rows, int cols) {
+Matrix2D Matrix2D::zeros(unsigned short rows, unsigned short cols) {
 	vector* __d = new vector(rows * cols, 0.0);
 	return Matrix2D(__d, rows, cols);
 }
 
-Matrix2D Matrix2D::identity(int _l) {
+Matrix2D Matrix2D::identity(unsigned short _l) {
 	vector tmpvec(_l, 1.);
 	return diagonal(tmpvec);
 }
 
 Matrix2D Matrix2D::diagonal(vector& _diag) {
 	vector* data = new vector;
-	int _l = _diag.size();
-	int arr_size =  _l * _l;
+	unsigned short _l = _diag.size();
+	unsigned short arr_size =  _l * _l;
 	data->reserve(arr_size);
-	int diag_idx = 0;
-	int val_arr_idx = 0;
+	unsigned short diag_idx = 0;
+	unsigned short val_arr_idx = 0;
 	for (unsigned i = 0; i < arr_size; i++) {
 		if (i == diag_idx) {
 			data->emplace_back(_diag[val_arr_idx]);
@@ -71,16 +71,16 @@ Matrix2D Matrix2D::diagonal(vector& _diag) {
 
 // PROPERTIES
 
-const int Matrix2D::size() const {
+const unsigned short Matrix2D::size() const {
 
 	return __data->size();
 }
 
-const int Matrix2D::numRows() const {
+const unsigned short Matrix2D::numRows() const {
 	return __rows;
 }
 
-const int Matrix2D::numCols() const {
+const unsigned short Matrix2D::numCols() const {
 	return __cols;
 }
 
@@ -107,7 +107,7 @@ const bool Matrix2D::isSquare() const {
 }
 
 const bool Matrix2D::isDiagonal() const {
-	int coeff = 0;
+	unsigned short coeff = 0;
 	for (int i = 0; i < size(); i++) {
 		if (i != coeff) {
 			if (__data->at(i) != 0.) {
@@ -122,7 +122,7 @@ const bool Matrix2D::isDiagonal() const {
 }
 const bool Matrix2D::isIdentity() const {
 	if (isSquare()) {
-		int coeff = 0;
+		unsigned short coeff = 0;
 		for (int i = 0; i < size(); i++) {
 			if (i != coeff) {
 				if (__data->at(i) != 0.) {
@@ -272,7 +272,7 @@ int Matrix2D::nextIdx(int current_idx, int last_base) {
 	return current_idx + __rows;
 }
 
-void Matrix2D::reshape(int rows, int columns) {
+void Matrix2D::reshape(unsigned short rows, unsigned short columns) {
 	if (isValidReshape(rows, columns)) {
 		__rows = rows;
 		__cols = columns;
@@ -286,11 +286,11 @@ Matrix2D Matrix2D::shift(double coeff) {
 	}
 }
 
-double Matrix2D::valueAt(int row, int column) {
+double Matrix2D::valueAt(unsigned short row, unsigned short column) {
 	return realValueAt(row - 1, column - 1);
 }
 
-double Matrix2D::realValueAt(int row_index, int column_index) {
+double Matrix2D::realValueAt(unsigned short row_index, unsigned short column_index) {
 	if (row_index < __rows and column_index < __cols) {
 		unsigned idx = (row_index * __cols) + column_index;
 		return __data->at(idx);
@@ -298,14 +298,14 @@ double Matrix2D::realValueAt(int row_index, int column_index) {
 	throw matrixOutOfRange(*this, row_index, column_index);
 }
 
-const vector Matrix2D::getRow(int row_number) const {
+const vector Matrix2D::getRow(unsigned short row_number) const {
 	return getRealRow(row_number - 1);
 }
 
-const vector Matrix2D::getRealRow(int row_index) const {
+const vector Matrix2D::getRealRow(unsigned short row_index) const {
 	if (0 <= row_index and row_index < __rows) {
-		int idx = row_index * __cols;
-		int end_idx = idx + __cols;
+		unsigned short idx = row_index * __cols;
+		unsigned short end_idx = idx + __cols;
 		vector row;
 		row.reserve(__cols);
 		for (; idx < end_idx; idx++) {
@@ -318,16 +318,16 @@ const vector Matrix2D::getRealRow(int row_index) const {
 	}
 }
 
-const vector Matrix2D::getColumn(int column_number) const {
+const vector Matrix2D::getColumn(unsigned short column_number) const {
 	return getRealColumn(column_number - 1);
 
 }
 
-const vector Matrix2D::getRealColumn(int column_idx) const {
+const vector Matrix2D::getRealColumn(unsigned short column_idx) const {
 	if (0 <= column_idx and column_idx < __cols) {
 		vector col;
 		col.reserve(__rows);
-		for (int i = column_idx; i < size(); i += __cols) {
+		for (unsigned short i = column_idx; i < size(); i += __cols) {
 			col.emplace_back(__data->at(i));
 		}
 		return col;
@@ -408,7 +408,7 @@ Matrix2D Matrix2D::horizontalBroadcastAddition(vector& vec, bool in_place) {
 
 // CHECKS
 
-bool Matrix2D::areValidParams(vector* data, int rows, int cols) {
+bool Matrix2D::areValidParams(vector* data, unsigned short rows, unsigned short cols) {
 	return (rows > 0) && (cols > 0) && (data->size() == rows * cols);
 }
 
@@ -428,7 +428,7 @@ bool Matrix2D::isSameShapeAs(Matrix2D& mx) {
 		and (__rows == mx.__rows);
 }
 
-bool Matrix2D::isValidReshape(int rows, int columns) {
+bool Matrix2D::isValidReshape(unsigned short rows, unsigned short columns) {
 	return rows * columns == size();
 }
